move tests to standard pkg/test folder, enable conftest options with soft-import, update readme and makefile
