+++ test_libs/pyspec/eth2spec/test/block_processing/test_process_block_header.py
@@ -6,12 +6,12 @@ from eth2spec.phase0.spec import (
     advance_slot,
     process_block_header,
 )
-from eth2spec.testing.helpers import (
+from eth2spec.test.helpers import (
     build_empty_block_for_next_slot,
     next_slot,
 )
 
-from eth2spec.testing.context import spec_state_test, expect_assertion_error
+from eth2spec.test.context import spec_state_test, expect_assertion_error
 
 
 def prepare_state_for_header_processing(state):
