+++ test_libs/pyspec/eth2spec/test/block_processing/test_process_deposit.py
@@ -4,7 +4,7 @@ from eth2spec.phase0.spec import (
     ZERO_HASH,
     process_deposit,
 )
-from eth2spec.testing.helpers import (
+from eth2spec.test.helpers import (
     get_balance,
     build_deposit,
     prepare_state_and_deposit,
@@ -12,7 +12,7 @@ from eth2spec.testing.helpers import (
     pubkeys,
 )
 
-from eth2spec.testing.context import spec_state_test, expect_assertion_error
+from eth2spec.test.context import spec_state_test, expect_assertion_error
 
 
 def run_deposit_processing(state, deposit, validator_index, valid=True):
