+++ test_libs/pyspec/eth2spec/test/block_processing/test_process_transfer.py
@@ -6,12 +6,12 @@ from eth2spec.phase0.spec import (
     get_current_epoch,
     process_transfer,
 )
-from eth2spec.testing.helpers import (
+from eth2spec.test.helpers import (
     get_valid_transfer,
     next_epoch,
 )
 
-from eth2spec.testing.context import spec_state_test, expect_assertion_error
+from eth2spec.test.context import spec_state_test, expect_assertion_error
 
 
 def run_transfer_processing(state, transfer, valid=True):
