+++ test_libs/pyspec/eth2spec/test/block_processing/test_process_voluntary_exit.py
@@ -6,12 +6,12 @@ from eth2spec.phase0.spec import (
     get_current_epoch,
     process_voluntary_exit,
 )
-from eth2spec.testing.helpers import (
+from eth2spec.test.helpers import (
     build_voluntary_exit,
     pubkey_to_privkey,
 )
 
-from eth2spec.testing.context import spec_state_test, expect_assertion_error
+from eth2spec.test.context import spec_state_test, expect_assertion_error
 
 
 def run_voluntary_exit_processing(state, voluntary_exit, valid=True):
