+++ test_libs/pyspec/eth2spec/test/epoch_processing/test_process_registry_updates.py
@@ -4,11 +4,11 @@ from eth2spec.phase0.spec import (
     get_current_epoch,
     is_active_validator,
 )
-from eth2spec.testing.helpers import (
+from eth2spec.test.helpers import (
     next_epoch,
 )
 
-from eth2spec.testing.context import spec_state_test
+from eth2spec.test.context import spec_state_test
 
 
 @spec_state_test
